requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
base58==2.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-ipfs-storage==0.0.4
django-rest-knox==4.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
filelock==3.8.2
idna==3.4
install==1.3.5
ipfsapi==0.4.4
ipfshttpclient==0.6.0
kombu==5.2.4
multiaddr==0.0.9
netaddr==0.8.0
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==4.21.12
pycodestyle==2.10.0
pycparser==2.21
PythonVideoConverter==1.0.3
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
varint==1.0.2
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.5
